Move general logic to the `list-store` package
