+++ packages/list-store/src/index.ts
@@ -1,2 +1,6 @@
 export * from './store';
 export * from './sort';
+// Explicitly reexported client components.
+// https://github.com/vercel/next.js/issues/41940#issuecomment-1480885131
+export {ClientStore} from './client-store';
+export * from './hooks';
