+++ packages/list-store/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "tsconfig/base.json",
+  "extends": "tsconfig/react-library.json",
   "compilerOptions": {
     "rootDir": "./src",
     "outDir": "./build"
